requirements.txt
feedparser==6.0.11
python-dateutil==2.9.0.post0
openai>=1.30.0
requests>=2.32.3
jinja2>=3.1.4
tenacity>=8.2.3